-add subproject -add test -fix gitignore
