+++ gradle.properties
@@ -1,1 +1,3 @@
 kotlin.code.style=official
+
+kotlinVersion=2.0.0
