+++ m1l1-first/src/main/kotlin/Main.kt
@@ -1,0 +1,5 @@
+package ru.otus.otuskotlin.marketplace
+
+fun main() {
+    println("Hello, World!")
+}
