+++ m1l1-first/src/test/kotlin/FirstTest.kt
@@ -1,0 +1,9 @@
+import kotlin.test.Test
+import kotlin.test.assertEquals
+
+class FirstTest {
+    @Test
+    fun firstTest() {
+        assertEquals(3, 1 + 2)
+    }
+}
