+++ settings.gradle.kts
@@ -1,5 +1,13 @@
+pluginManagement {
+    val kotlinVersion: String by settings
+    plugins {
+        kotlin("jvm") version kotlinVersion
+    }
+}
+
 plugins {
     id("org.gradle.toolchains.foojay-resolver-convention") version "0.5.0"
 }
 rootProject.name = "otus-marketplace"
 
+include("m1l1-first")
